Add public site as subtree, automatically deploy it (#60)
I'm using subtree here for the first time. Probably could be using it better. But the end result is that the repo at https://github.com/umn-terms/umn-terms.github.io is now subtree'd into this repo under `./site`. This brought along all of that repo's commits, which is where some of the commit noise comes from.

The juicy bit of this commit is this deploy script

https://github.umn.edu/asrweb/at_the_sound_of_the_tone/compare/subtree_public_site?expand=1#diff-639aff00aa4ee8f19030e8c375bd6652

Which does the following

- stashes changes
- switches to master
- deploys master
- updates the site with the contents of this repo's README
- deploys the site